# Conflicts:
#	src/main/java/io/ebeaninternal/server/deploy/AssocOneHelpRefInherit.java
#	src/main/java/io/ebeaninternal/server/query/SqlTreeNodeBean.java
#	src/test/java/org/tests/model/inheritmany/TestMediaInheritanceJoinToMany.java

+++ src/test/java/io/ebeaninternal/server/cache/CacheBeanDataTest.java
@@ -37,7 +37,7 @@ public class CacheBeanDataTest extends BaseTestCase {
     c.setSmallnote("somenote");
 
     Address billingAddress = new Address();
-    billingAddress.setId((short) 12);
+    billingAddress.setId(12);
     billingAddress.setCity("Auckland");
     billingAddress.setCountry(server.getReference(Country.class, "NZ"));
     billingAddress.setLine1("92 Someplace Else");
